[scanner] test: split 6 oversized hooks test files
Splits useMCS.test.ts (1088 lines) into three focused modules:
- useMCS.status.test.ts   — useMCSStatus tests
- useMCS.exports.test.ts  — useServiceExports + useServiceExport tests
- useMCS.imports.test.ts  — useServiceImports + useServiceImport tests

All existing tests preserved; no comment-only stubs; no generated
artifacts committed. The split files for issues #20989, #20992,
#20993, #20994 and #20995 are already present on main from a prior
merge; this PR closes all six issues by reference.
